# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Temporary Group Maker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface CreateUserBody {
  /**
   * @minLength 2
   * @maxLength 50
   */
  name: string;
  email: string;
  avatarUrl?: string | null;
}

export type GroupPurpose = (typeof GroupPurpose)[keyof typeof GroupPurpose];

export const GroupPurpose = {
  event: "event",
  project: "project",
  decision: "decision",
  brainstorm: "brainstorm",
  other: "other",
} as const;

export type GroupStatus = (typeof GroupStatus)[keyof typeof GroupStatus];

export const GroupStatus = {
  ACTIVE: "ACTIVE",
  EXPIRED: "EXPIRED",
  CLOSED: "CLOSED",
  ARCHIVED: "ARCHIVED",
} as const;

export interface Group {
  id: string;
  name: string;
  purpose: GroupPurpose;
  createdAt: string;
  deadline: string;
  status: GroupStatus;
  adminId: string;
  parentGroupId?: string | null;
}

export type GroupWithMetaUrgency =
  (typeof GroupWithMetaUrgency)[keyof typeof GroupWithMetaUrgency];

export const GroupWithMetaUrgency = {
  safe: "safe",
  warning: "warning",
  critical: "critical",
  expired: "expired",
} as const;

export type GroupWithMeta = Group & {
  memberCount: number;
  messageCount: number;
  adminName: string;
  urgency: GroupWithMetaUrgency;
  hoursUntilDeadline?: number | null;
  hasSummary: boolean;
  childGroupCount: number;
};

export type CreateGroupBodyPurpose =
  (typeof CreateGroupBodyPurpose)[keyof typeof CreateGroupBodyPurpose];

export const CreateGroupBodyPurpose = {
  event: "event",
  project: "project",
  decision: "decision",
  brainstorm: "brainstorm",
  other: "other",
} as const;

export interface CreateGroupBody {
  /**
   * @minLength 3
   * @maxLength 50
   */
  name: string;
  purpose: CreateGroupBodyPurpose;
  deadline: string;
  memberIds?: string[];
}

export type UpdateGroupBodyStatus =
  (typeof UpdateGroupBodyStatus)[keyof typeof UpdateGroupBodyStatus];

export const UpdateGroupBodyStatus = {
  ACTIVE: "ACTIVE",
  EXPIRED: "EXPIRED",
  CLOSED: "CLOSED",
  ARCHIVED: "ARCHIVED",
} as const;

export interface UpdateGroupBody {
  /**
   * @minLength 3
   * @maxLength 50
   */
  name?: string;
  status?: UpdateGroupBodyStatus;
}

export interface ExtendGroupBody {
  newDeadline: string;
  reason?: string | null;
}

export interface BranchSpec {
  topic: string;
  memberIds: string[];
  durationDays: number;
}

export interface BranchGroupBody {
  /** @minItems 1 */
  branches: BranchSpec[];
}

export interface Message {
  id: string;
  groupId: string;
  userId: string;
  content: string;
  createdAt: string;
  userName: string;
  userAvatarUrl?: string | null;
}

export interface SendMessageBody {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
  userId: string;
}

export type GroupMemberWithUserRole =
  (typeof GroupMemberWithUserRole)[keyof typeof GroupMemberWithUserRole];

export const GroupMemberWithUserRole = {
  ADMIN: "ADMIN",
  PARTICIPANT: "PARTICIPANT",
} as const;

export interface GroupMemberWithUser {
  groupId: string;
  userId: string;
  role: GroupMemberWithUserRole;
  joinedAt: string;
  user: User;
}

export type AddGroupMemberBodyRole =
  (typeof AddGroupMemberBodyRole)[keyof typeof AddGroupMemberBodyRole];

export const AddGroupMemberBodyRole = {
  ADMIN: "ADMIN",
  PARTICIPANT: "PARTICIPANT",
} as const;

export interface AddGroupMemberBody {
  userId: string;
  role?: AddGroupMemberBodyRole;
}

export type GroupSummarySuggestedNextAction =
  | (typeof GroupSummarySuggestedNextAction)[keyof typeof GroupSummarySuggestedNextAction]
  | null;

export const GroupSummarySuggestedNextAction = {
  archive: "archive",
  extend: "extend",
  branch: "branch",
  task_board: "task_board",
  decision_board: "decision_board",
} as const;

export interface GroupSummary {
  id: string;
  groupId: string;
  generatedAt: string;
  keyDecisions: string[];
  actionItems: string[];
  participantCount: number;
  suggestedNextAction?: GroupSummarySuggestedNextAction;
}

export interface PurposeCount {
  purpose: string;
  count: number;
}

export interface DashboardStats {
  activeGroups: number;
  expiredGroups: number;
  closedGroups: number;
  archivedGroups: number;
  totalMessages: number;
  groupsByPurpose: PurposeCount[];
}

export type ActivityItemUrgency =
  (typeof ActivityItemUrgency)[keyof typeof ActivityItemUrgency];

export const ActivityItemUrgency = {
  safe: "safe",
  warning: "warning",
  critical: "critical",
  expired: "expired",
} as const;

export interface ActivityItem {
  groupId: string;
  groupName: string;
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  lastMessageUser?: string | null;
  urgency: ActivityItemUrgency;
}

export type ListUsersParams = {
  search?: string;
};

export type ListGroupsParams = {
  status?: ListGroupsStatus;
};

export type ListGroupsStatus =
  (typeof ListGroupsStatus)[keyof typeof ListGroupsStatus];

export const ListGroupsStatus = {
  ACTIVE: "ACTIVE",
  EXPIRED: "EXPIRED",
  CLOSED: "CLOSED",
  ARCHIVED: "ARCHIVED",
} as const;

export type ListMessagesParams = {
  limit?: number;
  before?: string;
};