# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Temporary Group Maker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List groups for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, EXPIRED, CLOSED, ARCHIVED]
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupWithMeta"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new temporary group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupWithMeta"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group by ID
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupWithMeta"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group (extend deadline, change status)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupBody"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupWithMeta"
  /groups/{groupId}/close:
    post:
      operationId: closeGroup
      tags: [groups]
      summary: Close a group permanently (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Closed group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupWithMeta"
  /groups/{groupId}/extend:
    post:
      operationId: extendGroup
      tags: [groups]
      summary: Extend group deadline (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtendGroupBody"
      responses:
        "200":
          description: Extended group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupWithMeta"
  /groups/{groupId}/branch:
    post:
      operationId: branchGroup
      tags: [groups]
      summary: Create branch subgroups from a parent group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchGroupBody"
      responses:
        "201":
          description: Created branch groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupWithMeta"
  /groups/{groupId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      tags: [members]
      summary: List members of a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMemberWithUser"
    post:
      operationId: addGroupMember
      tags: [members]
      summary: Add a member to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGroupMemberBody"
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMemberWithUser"
  /groups/{groupId}/members/{userId}:
    delete:
      operationId: removeGroupMember
      tags: [members]
      summary: Remove a member from a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member removed
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        purpose:
          type: string
          enum: [event, project, decision, brainstorm, other]
        createdAt:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [ACTIVE, EXPIRED, CLOSED, ARCHIVED]
        adminId:
          type: string
        parentGroupId:
          type: string
          nullable: true
      required:
        - id
        - name
        - purpose
        - createdAt
        - deadline
        - status
        - adminId
    GroupWithMeta:
      allOf:
        - $ref: "#/components/schemas/Group"
        - type: object
          properties:
            memberCount:
              type: integer
            messageCount:
              type: integer
            adminName:
              type: string
            urgency:
              type: string
              enum: [safe, warning, critical, expired]
            hoursUntilDeadline:
              type: number
              nullable: true
            hasSummary:
              type: boolean
            childGroupCount:
              type: integer
          required:
            - memberCount
            - messageCount
            - adminName
            - urgency
            - hasSummary
            - childGroupCount
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
        purpose:
          type: string
          enum: [event, project, decision, brainstorm, other]
        deadline:
          type: string
          format: date-time
        memberIds:
          type: array
          items:
            type: string
      required:
        - name
        - purpose
        - deadline
    UpdateGroupBody:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
        status:
          type: string
          enum: [ACTIVE, EXPIRED, CLOSED, ARCHIVED]
    ExtendGroupBody:
      type: object
      properties:
        newDeadline:
          type: string
          format: date-time
        reason:
          type: string
          nullable: true
      required:
        - newDeadline
    BranchGroupBody:
      type: object
      properties:
        branches:
          type: array
          items:
            $ref: "#/components/schemas/BranchSpec"
          minItems: 1
      required:
        - branches
    BranchSpec:
      type: object
      properties:
        topic:
          type: string
        memberIds:
          type: array
          items:
            type: string
        durationDays:
          type: integer
      required:
        - topic
        - memberIds
        - durationDays
    Message:
      type: object
      properties:
        id:
          type: string
        groupId:
          type: string
        userId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        userName:
          type: string
        userAvatarUrl:
          type: string
          nullable: true
      required:
        - id
        - groupId
        - userId
        - content
        - createdAt
        - userName
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        userId:
          type: string
      required:
        - content
        - userId
    GroupMemberWithUser:
      type: object
      properties:
        groupId:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [ADMIN, PARTICIPANT]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required:
        - groupId
        - userId
        - role
        - joinedAt
        - user
    AddGroupMemberBody:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [ADMIN, PARTICIPANT]
          default: PARTICIPANT
      required:
        - userId
...