# lib/api-zod/src/generated/types/createGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Temporary Group Maker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGroupBodyPurpose } from "./createGroupBodyPurpose";

export interface CreateGroupBody {
  /**
   * @minLength 3
   * @maxLength 50
   */
  name: string;
  purpose: CreateGroupBodyPurpose;
  deadline: Date;
  memberIds?: string[];
}