# lib/api-zod/src/generated/types/extendGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Temporary Group Maker API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtendGroupBody {
  newDeadline: Date;
  reason?: string | null;
}