# lib/api-zod/src/generated/types/groupWithMeta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Temporary Group Maker API
 * OpenAPI spec version: 0.1.0
 */
import type { Group } from "./group";
import type { GroupWithMetaUrgency } from "./groupWithMetaUrgency";

export type GroupWithMeta = Group & {
  memberCount: number;
  messageCount: number;
  adminName: string;
  urgency: GroupWithMetaUrgency;
  hoursUntilDeadline?: number | null;
  hasSummary: boolean;
  childGroupCount: number;
};